refactor: extract remapWordTextsPreservingTiming to break import cycle
